fix: revert nonTransitiveRClass and fix compilation errors
- Revert nonTransitiveRClass=true back to false (was causing R class resolution failures)
- Add missing imports for DeleteNonLibraryManga and RemoveResettedHistory in DomainModule
- Fix Map<String?, Int> vs Map<String, Int> type mismatch in LibraryUpdateJob
- Fix Int vs Long type mismatch for remote_id in SuwayomiApi

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/suwayomi/SuwayomiApi.kt
@@ -70,7 +70,7 @@ class SuwayomiApi(private val trackId: Long) {
 
         result.map { manga ->
             TrackSearch.create(trackId).apply {
-                remote_id = manga.id
+                remote_id = manga.id.toLong()
                 title = manga.title
                 cover_url = manga.thumbnailUrl?.let { "$baseUrl/$it" } ?: ""
                 summary = manga.description.orEmpty()
